// src/app/models/global/globalGeneric.ts
export interface IAttachment {
  id: number;
  moduleId?: string;
  fileName?: string;
  fileType?: string;
  docUrl?: string;
  itemId?: number;
  file64Bytes?: string;
  fileSize?: number;
  docGUID?: string;
  createdOn?: string;
  createdBy?: string;
  updatedOn?: string;
  updatedBy?: string;
  isActive?: boolean;
  formFile?: any;
  formData?: any;
  folderURL?: string;
  createdByEmailAddress?: string;
  draftId?: string; //Internal ID for managing draft attachments which do not have module id
  moduleTypeId?: number | undefined;
  isAllowDelete?: boolean;
  chunkFileReferenceGuid?: string;
  chunkFileSize?: number;
  chunkFileTempPath?: string;
  sectionId?: number;
  componentId?: number;
  storageServer?: string;
  contentType?: number;
  sharing?: number;
  overwriteExistingFile?: boolean;
  folderId?: number;
  currentShareFileIndex?: number;
  totalNumberOfUploadedFiles?: number;
  docTitle?: string;
}